pyproject.toml
[tool.poetry]
name = "telegram-notes-bot"
version = "0.1.0"
description = "Telegram bot for saving text messages to markdown notes"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"